Re-organize the module layout
Change to private types that have no use.
Move to make import more convenient.
Fix inconsistent type names.
