+++ examples/basic.rs
@@ -1,6 +1,6 @@
 use std::sync::Arc;
 
-use aws_gamelift_server_sdk_rs::{Api, LogParameters, ProcessParameters, ServerParameters};
+use aws_gamelift_server_sdk_rs::*;
 
 static API: tokio::sync::OnceCell<Api> = tokio::sync::OnceCell::const_new();
