+++ src/model/mod.rs
@@ -1,15 +1,12 @@
 mod data;
-pub mod message;
+mod error;
+pub(crate) mod message;
 pub mod protocol;
 pub mod request;
 pub mod responce_result;
 
 pub use data::*;
-
+pub use error::Error;
 pub use message::{
     CreateGameSessionMessage as GameSession, UpdateGameSessionMessage as UpdateGameSession,
-};
-pub use request::{
-    DescribePlayerSessionsRequest, GetFleetRoleCredentialsRequest, StartMatchBackfillRequest,
-    StopMatchBackfillRequest,
 };
